♻️ [Refactor] Restructure files by dependency
