+++ Brewnary/Brewnary/Application/UseCase/BeanUseCase/FlavorService.swift
@@ -1,0 +1,18 @@
+//
+//  FlavorService.swift
+//  Brewnary
+//
+//  Created by 박진홍 on 5/9/25.
+//
+
+final class FlavorService {
+    private let repository: FlavorRepository
+    
+    init(repository: FlavorRepository) {
+        self.repository = repository
+    }
+    
+    func createFlavor(_ name: String) {
+        
+    }
+}
