+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/BeanEntity+CoreDataClass.swift
