+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/BeanEntity+CoreDataProperties.swift
