+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/FlavorEntity+CoreDataClass.swift
