+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/FlavorEntity+CoreDataProperties.swift
