+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/RoastLevelEntity+CoreDataClass.swift
