+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/RoastLevelEntity+CoreDataProperties.swift
