+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/RoasterEntity+CoreDataClass.swift
