+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/RoasterEntity+CoreDataProperties.swift
