+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/StorageTypeEntity+CoreDataClass.swift
