+++ Brewnary/Brewnary/Data/CoreDataModel/BeanEntity/StorageTypeEntity+CoreDataProperties.swift
