+++ Brewnary/Brewnary/Data/Support/MapperHelper.swift
