+++ Brewnary/Brewnary/Data/Support/MapperType.swift
