+++ Brewnary/Brewnary/Domain/Error/CoreDataError.swift
