+++ Brewnary/Brewnary/Domain/Error/MappingError.swift
