+++ Brewnary/Info.plist
